+++ pkgs/unified_analytics/example/unified_analytics_example.dart
@@ -13,7 +13,11 @@ final Analytics analytics = Analytics.development(
   tool: DashTool.flutterTool,
   flutterChannel: 'ey-test-channel',
   flutterVersion: 'Flutter 3.6.0-7.0.pre.47',
+  clientIde: 'VSCode',
   dartVersion: 'Dart 2.19.0',
+  // This can be set to true while testing to validate
+  // against GA4 usage limitations (character limits, etc.)
+  enableAsserts: false,
 );
 
 // Timing a process and sending the event