+++ pkgs/unified_analytics/lib/src/user_property.dart
@@ -17,6 +17,9 @@ class UserProperty {
   final String? flutterVersion;
   final String dartVersion;
   final String tool;
+  final String hostOsVersion;
+  final String locale;
+  final String? clientIde;
 
   /// This class is intended to capture all of the user's
   /// metadata when the class gets initialized as well as collecting
@@ -28,6 +31,9 @@ class UserProperty {
     required this.flutterVersion,
     required this.dartVersion,
     required this.tool,
+    required this.hostOsVersion,
+    required this.locale,
+    required this.clientIde,
   });
 
   /// This method will take the data in this class and convert it into
@@ -60,5 +66,8 @@ class UserProperty {
         'analytics_pkg_version': kPackageVersion,
         'tool': tool,
         'local_time': formatDateTime(clock.now()),
+        'host_os_version': hostOsVersion,
+        'locale': locale,
+        'client_ide': clientIde,
       };
 }