+++ pkgs/unified_analytics/test/unified_analytics_test.dart
@@ -44,6 +44,9 @@ void main() {
   const flutterVersion = 'flutterVersion';
   const dartVersion = 'dartVersion';
   const platform = DevicePlatform.macos;
+  const hostOsVersion = 'Version 14.1 (Build 23B74)';
+  const locale = 'en';
+  const clientIde = 'VSCode';
 
   final testEvent = Event.hotReloadTime(timeMs: 50);
 
@@ -89,6 +92,7 @@ void main() {
       dartVersion: dartVersion,
       fs: fs,
       platform: platform,
+      clientIde: clientIde,
     );
     analytics.clientShowedMessage();
 
@@ -115,6 +119,9 @@ void main() {
       flutterVersion: flutterVersion,
       dartVersion: dartVersion,
       tool: initialTool.label,
+      hostOsVersion: hostOsVersion,
+      locale: locale,
+      clientIde: clientIde,
     );
   });
 
@@ -594,6 +601,9 @@ ${initialTool.label}=$dateStamp,$toolsMessageVersion
       'analytics_pkg_version',
       'tool',
       'local_time',
+      'host_os_version',
+      'locale',
+      'client_ide',
     ];
     expect(analytics.userPropertyMap.keys.length, userPropertyKeys.length,
         reason: 'There should only be ${userPropertyKeys.length} keys');