+++ pkgs/unified_analytics/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 5.3.0
+
+- User property "host_os_version" added to provide detail version information about the host
+- User property "locale" added to provide language related information
+- User property "client_ide" (optional) added to provide the IDE used by the Dash tool using this package, if applicable
+
 ## 5.2.0
 
 - Added the `Event.hotRunnerInfo` constructor